allNoteDone

+++ app/Http/Controllers/Api/ForgetPasswordController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Mail\SendCodeResetPassword;
 use App\Models\ResetPasswordCode;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Mail;
 
@@ -23,31 +24,40 @@ class ForgetPasswordController extends Controller
             'email' => 'required',
         ]);
 
-        // Delete all old code that user send before.
-        ResetPasswordCode::where('email', $request->email)->delete();
+        $isEmail_Exist = User::where('email', $request->email)->exists();
 
-        // Generate random code
-        $gen_code = mt_rand(100000, 999999);
+        if ($isEmail_Exist) {
+            // Delete all old code that user send before.
+            ResetPasswordCode::where('email', $request->email)->delete();
 
-        // save info to DB
-        $codeData = new ResetPasswordCode();
-        $codeData->email = $request->email;
-        $codeData->code = $gen_code;
-        $codeData->created_at = now();
-        $codeData->save();
+            // Generate random code
+            $gen_code = mt_rand(100000, 999999);
 
-        //Sent Mail with Code 
-        // Mail::to($request->email)->send(new SendCodeResetPassword($codeData->code));
+            // save info to DB
+            $codeData = new ResetPasswordCode();
+            $codeData->email = $request->email;
+            $codeData->code = $gen_code;
+            $codeData->created_at = now();
+            $codeData->save();
 
-        Mail::send('email.send-code-reset-password', ['code' => $gen_code], function ($message) use ($request) {
-            $message->to($request->email);
-            $message->subject('Reset Password Mail');
-        });
+            //Sent Mail with Code 
+            // Mail::to($request->email)->send(new SendCodeResetPassword($codeData->code));
 
-        //semd resposne
-        return response()->json([
-            'message' => "Reset code has been sent to your gmail",
-            'otpCode' => $gen_code
-        ], 200);
+            Mail::send('email.send-code-reset-password', ['code' => $gen_code], function ($message) use ($request) {
+                $message->to($request->email);
+                $message->subject('Reset Password Mail');
+            });
+
+            //semd resposne
+            return response()->json([
+                'message' => "Reset code has been sent to your gmail"
+            ], 200);
+        } else {
+            //semd resposne
+            return response()->json([
+                'message' => "email not found",
+                'response_code' => "404"
+            ], 404);
+        }
     }
 }
